Split bundle into app and vendor
The reason we do this is, presumably, the vendor code should not change
as often as our app code. Therefore, once the user visits our site the
first time, the vendor bundle will be in the browser cache and will not
have to be downloaded. The vendor bundle is often larger than the app
bundle so this is a big win.

The things are needed in our webpack config:
1. The entry [] becomes an object with keys for app and vendor.
2. Use the webpack CommonChunksPlugin

Our html file will need to have script tags for both vendor and app
bundles. The vendor bundle will include the webpack runtime so it needs
to be loaded before the app bundle.