Enable long term caching of assets
We setup code splitting of our bundle to attempt long term caching.
However, it does not work as intended because the vendor bundle, our
entry point, includes the chunk file hashes. Whenever we modify our app
code, the hash is going to change and the vendor bundle will have the
new hash. Thus the vendor hash will change too. To get around this we
configure webpack to extract the webpack runtime into a seperate
manifest.js file so the vendor does not have to change. The downside is
that this is an additional file that will have to be downloaded by the
browser. It is very small after minification and compression though.